requirements.txt
# Core
vanna==0.7.9
anthropic>=0.40.0
chromadb<1.0.0

# Embeddings
langchain-huggingface
sentence-transformers

# Utilities
python-dotenv
flask>=3.0.0
flask-cors
pandas
plotly
pytest
